Removed gFPS. Nobody needs FPS anyway. Its only a visual "gawd, noggit is slow". MPQFile: Now replacing / with \ to ensure it being correct, even if the path given was wrong. Noggit: Cleaned up the main loop a bit. Replaced some int with size_t to ensure 64 bit compatibility. Manager.h: Removed useless includes.

+++ src/Manager.h
@@ -1,9 +1,6 @@
 #ifndef MANAGER_H
 #define MANAGER_H
 
-#include <string> // std::string
-#include <algorithm> // std::transform()
-
 // base class for manager objects
 
 //! \todo Proxy objects / handles holding the name and the object.
@@ -11,7 +8,7 @@
 class ManagedItem 
 {
 private:
-  int _referenceCount;
+  size_t _referenceCount;
   
 public:
   explicit ManagedItem( )
